重构FFT乘法；新增带有缓存的单边变换乘法
