+++ src/lammp/lmmpn/mul.c
@@ -129,5 +129,5 @@ void lmmp_sqr_(mp_ptr restrict dst, mp_srcptr restrict numa, mp_size_t na) {
     else if (na < MUL_FFT_THRESHOLD)
         lmmp_sqr_toom4_(dst, numa, na);
     else
-        lmmp_mul_fft_(dst, numa, na, numa, na);
+        lmmp_sqr_fft_(dst, numa, na);
 }
